Refactor CEGUI ext
