+++ examples_tests/13.BRDF_Explorer/BRDFExplorerApp.cpp
@@ -26,7 +26,7 @@ SOFTWARE.
 */
 
 #include "BRDFExplorerApp.h"
-#include "../../ext/CEGUI/CEGUI.h"
+#include "../../ext/CEGUI/ExtCEGUI.h"
 
 namespace irr
 {
